changed some fields,methods and add new feat.

+++ src/main/java/com/example/devfka/demo/Service/AddressService.java
@@ -41,25 +41,25 @@ public class AddressService {
 	
 	
 	
-	public Address saveAddress (String identyNumber,String description) {
-			Person person=_personService.getByPersonIdentyNum(identyNumber);
-			if(person!=null) {
-				Address address=new Address();
-				address.setDescription(description);
-				address.setPerson(person);
-				address.setPerson(person);
-				addressRepo.save(address);
-				return address;
+	public Address saveAddress (int id,Address newAdress) {
+			Person person=_personService.getPersonById(id);
+			if(person!=null && addressRepo.findByPersonId(person.getId())==null) {
+					Address addNewAddress=new Address();
+					addNewAddress.setPerson(person);
+					addNewAddress.setDescription(newAdress.getDescription());
+					addressRepo.save(addNewAddress);
+					return addNewAddress;
 			}else {
 				return null;
 			}
 	}
 	
-	public Address updateAddress(String description,String identyNum) {
-		Person person=_personService.getByPersonIdentyNum(identyNum);
-		if(person!=null) {
+	public Address updateAddress(int id,Address newAdress) {
+		Person person=_personService.getPersonById(id);
+		if(person!=null && addressRepo.findByPersonId(person.getId())!=null) {
 			Address address=addressRepo.findByPersonId(person.getId());
-			address.setDescription(description);
+			address.setDescription(newAdress.getDescription());
+			address.setPerson(person);
 			addressRepo.save(address);
 			return address;
 		}else {
@@ -67,11 +67,12 @@ public class AddressService {
 		}
 	}
 	
-	public void deleteAddress(String identyNum) {
-		Person person=_personService.getByPersonIdentyNum(identyNum);
+	public void deleteAddress(int id) {
+		Person person=_personService.getPersonById(id);
+		if(person!=null && addressRepo.findByPersonId(person.getId())!=null) {
 			Address addressForDelete=addressRepo.findByPersonId(person.getId());
 			addressRepo.delete(addressForDelete);
-			
+		}	
 	}
